Add packages to Java-Kafka directories
