+++ java/common/src/main/java/io/strimzi/common/TracingInitializer.java
@@ -1,4 +1,4 @@
-/*
+package io.strimzi.common;/*
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
