+++ java/http/java-http-consumer/src/main/java/HttpConsumerConfig.java
@@ -1,3 +1,5 @@
+import io.strimzi.common.TracingSystem;
+
 /*
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
