+++ java/http/java-http-producer/src/main/java/HttpProducerConfig.java
@@ -3,6 +3,8 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
+import io.strimzi.common.TracingSystem;
+
 public class HttpProducerConfig {
 
     private static final String DEFAULT_HOSTNAME = "localhost";
