+++ java/kafka/streams/pom.xml
@@ -82,7 +82,7 @@
                             </filters>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>KafkaStreamsExample</mainClass>
+                                    <mainClass>io.strimzi.kafka.streams.KafkaStreamsExample</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
