+++ java/kafka/streams/src/main/java/io/strimzi/kafka/streams/TracingKafkaClientSupplier.java
@@ -3,6 +3,8 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
+package io.strimzi.kafka.streams;
+
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.instrumentation.kafkaclients.KafkaTelemetry;
 import org.apache.kafka.clients.admin.Admin;
